feat: 점수 시스템 및 로직 구현, 테스트 추가

+++ src/main/java/kb/health/controller/response/FollowResponse.java
@@ -9,8 +9,8 @@ import lombok.Getter;
 public class FollowResponse {
     private Long followId;  // Follow 엔티티의 PK
     private String userName;
-    private int dayScore;
-    private int baseScore;
+    private double dayScore;
+    private double baseScore;
     private String profileImageUrl;
 
     public static FollowResponse create(Member member) {
